requirements.txt
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
python-dotenv==1.0.1
PyJWT==2.10.1
psycopg2-binary==2.9.10
supabase==2.22.0
gunicorn==21.2.0
